agent/requirements.txt
msgpack-rpc-python
beanstalkc
msgpack
pycurl